modelkit: move safe_np_dump out of core/models/tensorflow_model
