+++ modelkit/utils/serialization.py
@@ -1,0 +1,11 @@
+import numpy as np
+
+
+def safe_np_dump(obj):
+    if isinstance(obj, np.integer):
+        return int(obj)
+    elif isinstance(obj, np.floating):
+        return float(obj)
+    elif isinstance(obj, np.ndarray):
+        return obj.tolist()
+    return obj.item()
